Rename MainComponent, remove index.js as is unneeded
