--- src/index.js
@@ -1,7 +1,0 @@
-import MainComponent from "./MainComponent.vue";
-
-export default {
-  install(Vue) {
-    Vue.component("bike-routes", MainComponent);
-  },
-};
